refactor(mobile): extract feature hooks from session route
- Extract useImageBase64: image-to-base64 conversion + mime resolution
- Extract useBookmarkHandler: bookmark state + toggle logic
- Extract useSessionRecovery: milestone hydration on resume + AppState marker writer
- session/index.tsx: 1445 → 1264 LOC; pure refactor, no behavior change
- Fix liveTranscriptMilestones prop type: readonly string[] → readonly CelebrationReason[]
